Adding a how to licence with different licences
Murray Cadzow suggested "The lesson content the work of <me/org> and is licensed under <licence of choice> and copyrighted by <me/org>, please attribute <me/org>. The lesson style/template was the work of The Carpentries <link to repo>" in the carpentries slack as a way to manage the CC-BY for lesson framework (Carpentries style) and original lesson content developed by him. Adding a how to seems a useful addition and would increase the practicable resusability in any case.